Remove duplicate BERT pretraining modules
